Create poll.js
You NEED to add the first three sections to have access to our firebase database. the rest is a format sample for adding data to it. Do some of your own research for specific forms when you switch between arrays, strings, etc. and how you need to structure your data.